Segments when throwing

+++ main/java/com/yyon/grapplinghook/controllers/airfrictionController.java
@@ -26,7 +26,7 @@ import com.yyon.grapplinghook.vec;
 public class airfrictionController extends grappleController {
 	public final double playermovementmult = 0.5;
 	
-	public airfrictionController(int arrowId, int entityId, World world, vec pos, int maxlen, int id) {
+	public airfrictionController(int arrowId, int entityId, World world, vec pos, double maxlen, int id) {
 		super(arrowId, entityId, world, pos, maxlen, id);
 	}
