+++ main/java/com/yyon/grapplinghook/controllers/enderController.java
@@ -24,7 +24,7 @@ import com.yyon.grapplinghook.vec;
 public class enderController extends grappleController {
 	public final double playermovementmult = 1;
 	
-	public enderController(int arrowId, int entityId, World world, vec pos, int maxlen, int id) {
+	public enderController(int arrowId, int entityId, World world, vec pos, double maxlen, int id) {
 		super(arrowId, entityId, world, pos, maxlen, id);
 	}
