+++ main/java/com/yyon/grapplinghook/controllers/hookControl.java
@@ -24,7 +24,7 @@ import com.yyon.grapplinghook.vec;
  */
 
 public class hookControl extends grappleController {
-	public hookControl(int arrowId, int entityId, World world, vec pos, int maxlen, int id) {
+	public hookControl(int arrowId, int entityId, World world, vec pos, double maxlen, int id) {
 		super(arrowId, entityId, world, pos, maxlen, id);
 	}
