+++ main/java/com/yyon/grapplinghook/controllers/magnetController.java
@@ -33,7 +33,7 @@ public class magnetController extends grappleController {
 	public int repelconf = 0;
 	public final double playermovementmult = 1.5;
 		
-	public magnetController(int arrowId, int entityId, World world, vec pos, int maxlen, int id, int repelconf) {
+	public magnetController(int arrowId, int entityId, World world, vec pos, double maxlen, int id, int repelconf) {
 		super(arrowId, entityId, world, pos, maxlen, id);
 		this.repelconf = repelconf;
 	}
